refactor(mobile): split chat screen into components; wire members button and channel kebab menu
