+++ mobile/components/chat/dates.ts
@@ -1,0 +1,28 @@
+// Day/time formatting helpers for the chat timeline.
+
+export function dayKey(ts: number): string {
+  return new Date(ts).toDateString();
+}
+
+export function dayLabel(ts: number): string {
+  const d = new Date(ts);
+  const today = new Date();
+  if (d.toDateString() === today.toDateString()) {
+    return "TODAY";
+  }
+  const yest = new Date(today);
+  yest.setDate(today.getDate() - 1);
+  if (d.toDateString() === yest.toDateString()) {
+    return "YESTERDAY";
+  }
+  return d
+    .toLocaleDateString(undefined, { month: "short", day: "numeric" })
+    .toUpperCase();
+}
+
+export function timeLabel(ts: number): string {
+  return new Date(ts).toLocaleTimeString(undefined, {
+    hour: "2-digit",
+    minute: "2-digit",
+  });
+}
